requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-login==0.6.3
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
eventlet==0.35.1
Pillow==10.4.0
psycopg2-binary==2.9.10
PyMySQL==1.1.1
